requirements.txt
requests==2.25.0
Flask==1.1.2
paho-mqtt==1.5.1